+++ regions/conftest.py
@@ -4,9 +4,7 @@
 # packagename.test
 
 import astropy.units as u
-import numpy as np
 import pytest
-from astropy.utils import minversion
 from astropy.wcs import WCS
 
 from regions._utils.optional_deps import HAS_GWCS
@@ -19,11 +17,6 @@ try:
     ASTROPY_HEADER = True
 except ImportError:
     ASTROPY_HEADER = False
-
-
-# do not remove until we drop support for NumPy < 2.0
-if minversion(np, '2.0.0.dev0+git20230726'):
-    np.set_printoptions(legacy='1.25')
 
 
 def pytest_configure(config):